refactor: application_detail 컴포넌트 기능단위 분리 및 폴더구조 개선
